Create CRUD.py
Clone the Project Repository

This lab requires multiple HTML interface files, which have been pre-created for you. You will need to clone the folder structure to the IDE interface using the following command in a terminal shell.

1
git clone https://github.com/ibm-developer-skills-network/obmnl-flask_assignment.git

When the command is successfully executed, the Project tab must have the folder structure as shown in the image. The root folder,  obmnl-flask_assignment should have the templates folder and a file app.py. The templates folder has all the required HTML files, edit.html, form.html, and transactions.html. Throughout this lab, you will implement the required functions in app.py.


Initial set up

In the app.py file, you need to import necessary modules from Flask and instantiate the Flask application.
For this lab, you will need to import the following functions from the flask library.

Flask - to instantiate the application
request - to process the GET and POST requests
url_for - to access the url for a given function using its decorator
redirect - to redirect access requests according to requirement
render_template - to render the html page
After importing the functions, instantiate the application to a variable app.

Next, let's create a list of sample transactions for testing purposes. You can assume that the transactions already exist on the interface when it is executed for the first time. Please note that this step is completely optional and does not affect the functionality you will develop in this lab. Add the code snippet as shown below to app.py.

The order in which you will develop the functions is as follows:
1. Read
2. Create
3. Update
4. Delete

The reason to implement Read before the other functions is to be able to redirect to the page with all transactions every time a new transaction is created, updated, or deleted. Therefore, the function to read the existing transactions must exist before the others are implemented.


Read Operation

To implement the Read operation, you need to implement a route that displays a list of all transactions. This route will handle GET requests, which are used to retrieve and display data in app.py.

The key steps to implement the Read operation are as follows:

Create a function named get_transactions that uses render_template to return an HTML template named transactions.html. This function should pass the transactions to the template for display.

Use the Flask @app.route decorator to map this function to the root (/) URL. This means that when a user visits the base URL of your application, Flask will execute the get_transactions function and return its result.


Create Operation

For the Create operation, you will implement a route that allows users to add new transactions. This will involve handling both GET and POST HTTP requests - GET for displaying the form to the user and POST for processing the form data sent by the user.

Here is the list of steps to implement the Create operation.

Create a function named add_transaction.

Use add as the decorator for this function. Make sure to pass both GET and POST as possible methods.

If the request method is GET, use the render_template function to display an HTML form using a template named form.html. This form will allow users to input data for a new transaction.

If the request method is POST, use request.form to extract the form data, create a new transaction, append it to the transactions list, and then use redirect and url_for to send the user back to the list of transactions.

The new transaction is passed on to the reading function in the following format.

Note: The statements outside the if case are, by default, the else case. The statements in the if case end with a return statement; hence only one of the two cases will run at a time.


Update Operation

For the Update operation, you need to implement a route that allows users to update existing transactions. You'll again handle both GET and POST HTTP requests - GET for displaying the current transaction data in a form, and POST for processing the updated data sent by the user.

Complete the following steps to implement the Update operation.

Create a function named edit_transaction that handles both GET and POST requests. This function should accept a parameter, transaction_id.

Decorate the function with @app.route and use the route string /edit/<int:transaction_id>. The <int:transaction_id> part in the URL is a placeholder for any integer. Flask will pass this integer to your function as the transaction_id argument.

If the request method is GET, find the transaction with the ID that matches transaction_id and use render_template to display a form pre-populated with the current data of the transaction using a template named edit.html.

If the request method is POST, use request.form to get the updated data, find the transaction with the ID that matches transaction_id and modify its data, then redirect the user back to the list of transactions.


Delete Operation

Finally, you need to implement a route that allows users to delete existing transactions.

Complete the following steps to implement the Delete operation.

Create a function named delete_transaction that takes a parameter, transaction_id.

Decorate the function with @app.route and use the route string /delete/<int:transaction_id>. The <int:transaction_id> part in the URL is a placeholder for any integer. Flask will pass this integer to your function as the transaction_id argument.

In the function body, find the transaction with the ID that matches transaction_id and remove it from the transactions list, then redirect the user back to the list of transactions.


Finishing Steps and Running the Application

Check if the current script is the main program (that is, it wasn't imported from another script) with the conditional if __name__ == "__main__":.

If the condition is true, call app.run(debug=True) to start the Flask development server with debug mode enabled. This will allow you to view detailed error messages in your browser if something goes wrong.

The code is now complete. Run the file app.py from a terminal shell using the command:

python3.11 app.py

By default, Flask launches the application on LocalHost:5000.

Launch the application by going to the Skills Network Library, going to Launch Application.

Enter 5000 in the port number and launch the application window.