Create GradioHello.py
Creates a gradio.Interface called demo. It wraps the greet function with a simple text-to-text user interface that you could interact with.

The gradio.Interface class is initialized with 3 required parameters:

fn: 
the function to wrap a UI around

inputs: 
which component(s) to use for the input (e.g. “text”, “image” or “audio”)

outputs: 
which component(s) to use for the output (e.g. “text”, “image” or “label”)

The last line demo.launch() launches a server to serve your demo.